moved util.py and predict.py to train.py: circular import issue
