--- recognition/StyleGAN2-OASIS_45711110/utils.py
@@ -1,52 +1,0 @@
-'''helper functions for model training'''
-
-import torch
-from config import w_dim, device, log_resolution
-import train
-
-def gradient_penalty(critic, real, fake,device="cpu"):
-    BATCH_SIZE, C, H, W = real.shape
-    beta = torch.rand((BATCH_SIZE, 1, 1, 1)).repeat(1, C, H, W).to(device)
-    interpolated_images = real * beta + fake.detach() * (1 - beta)
-    interpolated_images.requires_grad_(True)
-
-    # Calculate critic scores
-    mixed_scores = critic(interpolated_images)
- 
-    # Take the gradient of the scores with respect to the images
-    gradient = torch.autograd.grad(
-        inputs=interpolated_images,
-        outputs=mixed_scores,
-        grad_outputs=torch.ones_like(mixed_scores),
-        create_graph=True,
-        retain_graph=True,
-    )[0]
-    gradient = gradient.view(gradient.shape[0], -1)
-    gradient_norm = gradient.norm(2, dim=1)
-    gradient_penalty = torch.mean((gradient_norm - 1) ** 2)
-    return gradient_penalty
-
-# fetches w from mapping network
-def get_w(batch_size):
-
-    z = torch.randn(batch_size, w_dim).to(device)
-    w = train.mapping_network(z)
-    return w[None, :, :].expand(log_resolution, -1, -1)
-
-# Generates random noise
-def get_noise(batch_size):
-    
-    noise = []
-    resolution = 4
-
-    for i in range(log_resolution):
-        if i == 0:
-            n1 = None
-        else:
-            n1 = torch.randn(batch_size, 1, resolution, resolution, device=device)
-        n2 = torch.randn(batch_size, 1, resolution, resolution, device=device)
-
-        noise.append((n1, n2))
-        resolution *= 2
-
-    return noise
